+++ lib/std.ml
@@ -1,1 +1,2 @@
 module Bignum = Bignum0
+module Bigint = Bigint