+++ setup.ml
@@ -1,3 +1,7 @@
 (* OASIS_START *)
+#use "topfind";;
+#require "oasis.dynrun";;
+open OASISDynRun;;
+open OASISTypes;;
 (* OASIS_STOP *)
 let () = setup ()